refactor(test): `TestUtils`의 패키지 경로 수정
