+++ src/test/java/lotto/model/LottoResultCalculatorImplTest.java
@@ -1,11 +1,10 @@
 package lotto.model;
 
-import static lotto.util.TestUtils.*;
+import static lotto.test.util.TestUtils.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.util.TestUtils;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
