+++ src/test/java/lotto/model/WinningLottoTest.java
@@ -1,10 +1,9 @@
 package lotto.model;
 
-import static lotto.util.TestUtils.lottoNumbersOf;
+import static lotto.test.util.TestUtils.lottoNumbersOf;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.List;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
